# executor/docs/swagger.yaml
info:
  contact: {}
  description: This is the REST API for the executor
  termsOfService: http://swagger.io/terms/
  title: Executor API
  version: 0.1.0
paths:
  /images:
    post:
      consumes:
        - application/json
      description: Create a Docker image based on the provided parameters.
      parameters:
        - description: Name of the Docker image
          in: body
          name: ImageName
          required: true
          schema:
            type: string
        - description: Tag for the Docker image
          in: body
          name: Tag
          required: true
          schema:
            type: string
        - description: Content of the Dockerfile
          in: body
          name: Dockerfile
          required: true
          schema:
            type: string
      produces:
        - application/json
      responses:
        "200":
          description: '{success: true, message: "ImageName"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{success: false, message: "Error"}'
          schema:
            additionalProperties: true
            type: object
      summary: Create an image
      tags:
        - Images
  /images/delete:
    delete:
      consumes:
        - application/json
      description: Delete a Docker image by name and tag.
      parameters:
        - description: User ID
          in: header
          name: UserID
          required: true
          type: string
        - description: Name of the Docker image
          in: body
          name: ImageName
          required: true
          schema:
            type: string
        - description: Tag for the Docker image
          in: body
          name: Tag
          required: true
          schema:
            type: string
      produces:
        - application/json
      responses:
        "200":
          description: Successfully deleted the image
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete an image
      tags:
        - Images
  /items:
    get:
      consumes:
        - application/json
      description: Creates a sandbox using the specified environment
      operationId: get-item-list
      produces:
        - application/json
      responses:
        "200":
          description: OK
      summary: Create sandbox
swagger: "2.0"